requirements.txt
appdirs==1.4.3
bokeh==0.12.1
cycler==0.10.0
Django==1.11
django-crispy-forms==1.6.1
django-extensions==1.7.8
Jinja2==2.9.6
MarkupSafe==1.0
matplotlib==2.0.0
numpy==1.12.1
packaging==16.8
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
requests==2.13.0
six==1.10.0
tornado==4.4.3
